added ddp_spawn test
